tests: strictness leftovers in the lowering suite spell their conversions
Same migration class as 9a793a848c, in the 05 integration suite the
--lib baseline never covered: domain crossings at stores and call
arguments now spell 'as T in Domain' (or reorder so the domain operand
leads); the i8-in-Saturating local drops its pre-strictness trailing
'as i8' erasure.

+++ omega-rust/psi/pipeline/05_checked-trees-to-lowered-psi/tests/integer_policy_realization.rs
@@ -434,7 +434,7 @@ fn a_saturating_conversion_composes_on_signed_narrowing() {
         r#"
         data Main {}
         machine narrow(value: i16) {
-            let narrowed: i8 in Saturating = (value as i8 in Saturating) as i8;
+            let narrowed: i8 in Saturating = value as i8 in Saturating;
         }
         machine Main::main(value: i16) { narrow(value); }
     "#,
